feat: Add token to cookie if provided in query
This commit adds functionality to check if a token is provided in the query parameters of the request. If a token is found, it is stored in a cookie named 'token' with a maximum age of 900000 milliseconds and the 'httpOnly' flag set to true. This allows for easier token management and authentication.